requirements.txt
selenium>=4.0.0
webdriver-manager>=4.0.0
openpyxl>=3.0.0
pandas>=1.0.0
python-docx
pywin32
